Move TlfID to tlf
